+++ database/migrations/2019_10_11_131941_create_ordered_items_table.php
@@ -23,7 +23,6 @@ class CreateOrderedItemsTable extends Migration
             $table->integer('building');
             $table->integer('contractor');
             $table->integer('status');
-            $table->timestamps();
         });
     }