+++ database/seeds/DatabaseSeeder.php
@@ -15,10 +15,8 @@ class DatabaseSeeder extends Seeder
         $this->call(UsersRoleTableSeeder::class);
         $this->call(UsersDataTableSeeder::class);
         $this->call(DepartmentsTableSeeder::class);
-        $this->call(OrdersStatusTableSeeder::class);
         $this->call(BuildingsTableSeeder::class);
         $this->call(ContractorsTableSeeder::class);
-        $this->call(ItemsStatusTableSeeder::class);
         $this->call(UnitsTableSeeder::class);
     }
 }