+++ resources/lang/en/item_status.php
@@ -3,4 +3,5 @@
         'ordered' => 'Ordered',
         'not_ordered' => 'Not ordered',
         'delivered' => 'Delivered',
+        'unknown' => 'Unknown',
     ];