+++ resources/lang/en/print_order.php
@@ -1,0 +1,22 @@
+<?php
+    return [
+        /**
+         * Print order page
+         */
+        'print_order_number' => 'Num',
+        'print_order_date' => 'Order date',
+        'print_order_title' => 'ORDER',
+        'print_order_department' => 'Department',
+        'print_order_action' => 'orders the following materials',
+        'print_order_no' => 'No.',
+        'print_order_item_name' => 'Name',
+        'print_order_item_quantity' => 'Quantity',
+        'print_order_item_price' => 'Approx. price (per item)',
+        'print_order_item_building' => 'Building',
+        'print_order_item_sum_price' => 'Approx. price',
+        'print_order_sign_owner' => 'Owner signature',
+        'print_order_sign_manager' => 'Manager signature',
+        'print_order_consent' => 'Consent to purchase',
+        'print_order_num_items' => 'Number of ordered items',
+        'print_order_value' => 'Order value',
+    ];