+++ resources/lang/pl/print_order.php
@@ -1,0 +1,22 @@
+<?php
+    return [
+        /**
+         * Print order page
+         */
+        'print_order_number' => 'Nr.',
+        'print_order_date' => 'Data zamówienia',
+        'print_order_title' => 'ZAMÓWIENIE',
+        'print_order_department' => 'Dział',
+        'print_order_action' => 'zamawia następujące materiały',
+        'print_order_no' => 'Lp.',
+        'print_order_item_name' => 'Nazwa',
+        'print_order_item_quantity' => 'Ilość',
+        'print_order_item_price' => 'Cena przybl. (szt.)',
+        'print_order_item_building' => 'Budynek',
+        'print_order_item_sum_price' => 'Cena przybl.',
+        'print_order_sign_owner' => 'Podpis zamawiającego',
+        'print_order_sign_manager' => 'Podpis kierownika',
+        'print_order_consent' => 'Zgoda na zakup',
+        'print_order_num_items' => 'Liczba zamówionych rzeczy',
+        'print_order_value' => 'Wartość zamówienia',
+    ];